ENG-2037: Fix security issues

+++ src/main/java/org/entando/entando/aps/system/services/storage/FileBrowserService.java
@@ -15,7 +15,6 @@ package org.entando.entando.aps.system.services.storage;
 
 import com.agiletec.aps.util.FileTextReader;
 import java.security.SecureRandom;
-import org.apache.commons.lang.SystemUtils;
 import org.entando.entando.aps.system.exception.ResourceNotFoundException;
 import org.entando.entando.aps.system.exception.RestServerError;
 import org.entando.entando.aps.system.services.DtoBuilder;
@@ -32,9 +31,11 @@ import org.springframework.validation.BindingResult;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
+import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
 import java.util.List;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.entando.entando.web.common.exceptions.ValidationGenericException;
 import org.springframework.validation.BeanPropertyBindingResult;
@@ -88,23 +89,27 @@ public class FileBrowserService implements IFileBrowserService {
     @Override
     public byte[] getFileStream(String currentPath, Boolean protectedFolder) {
         this.checkResource(currentPath, "File", protectedFolder);
+        File tempDir = new File(FileTextReader.getTempDirectory());
         File tempFile = null;
         byte[] bytes = null;
         try {
             String[] sections = currentPath.split("/");
             InputStream stream = this.getStorageManager().getStream(currentPath, protectedFolder);
             tempFile = FileTextReader.createTempFile(new SecureRandom().nextInt(100) + sections[sections.length - 1], stream);
-            bytes = FileTextReader.fileToByteArray(tempFile);
+            bytes = FileTextReader.fileToByteArray(tempDir, tempFile);
         } catch (Throwable t) {
             logger.error("error extracting stream for path {} - type {}", currentPath, protectedFolder);
             throw new RestServerError("error extracting stream", t);
         } finally {
-            if (null != tempFile) {
-                boolean deleted = tempFile.delete();
-
-                if (!deleted) {
-                    logger.warn("Failed to delete temp file {}", tempFile.getAbsolutePath());
+            try {
+                if (null != tempFile && FileUtils.directoryContains(tempDir, tempFile)) {
+                    boolean deleted = tempFile.delete();
+                    if (!deleted) {
+                        logger.warn("Failed to delete temp file {}", tempFile.getAbsolutePath());
+                    }
                 }
+            } catch (IOException ex) {
+                logger.error("Error checking temp file", ex);
             }
         }
         return bytes;
